refactor: encapsulating editor logic layer to custom hook
