+++ src/components/editor/editor.tsx
@@ -15,7 +15,6 @@ import {
   Placeholder,
   handleCommandNavigation
 } from 'novel/extensions'
-import { useEffect, useMemo, useState } from 'react'
 import { defaultExtensions } from './extensions'
 import { ColorSelector } from './selectors/color-selector'
 import { LinkSelector } from './selectors/link-selector'
@@ -28,7 +27,8 @@ import { handleImageDrop, handleImagePaste } from 'novel/plugins'
 import { Toolbar } from '../rich-text-input/toolbar'
 import { uploadFn } from './image-upload'
 import { TextButtons } from './selectors/text-buttons'
-import { slashCommand, suggestionItems } from './slash-command'
+import { slashCommand } from './slash-command'
+import { useEditorState } from './use-editor-state'
 
 const extensions = [...defaultExtensions, slashCommand]
 
@@ -48,16 +48,18 @@ const NoteEditor = ({
   isSubmitting = false,
   variant = 'note'
 }: EditorProp) => {
-  const [openNode, setOpenNode] = useState(false)
-  const [openColor, setOpenColor] = useState(false)
-  const [openLink, setOpenLink] = useState(false)
-  const [editor, setEditor] = useState<Editor | null>(null)
-
-  useEffect(() => {
-    if (isSubmitting) {
-      editor?.commands.setContent(null)
-    }
-  }, [editor?.commands, isSubmitting])
+  const {
+    editor,
+    openColor,
+    openLink,
+    openNode,
+    setOpenColor,
+    setEditor,
+    setOpenLink,
+    setOpenNode,
+    filteredSuggestionItems,
+    hasH1TitleEnabled
+  } = useEditorState({ isSubmitting, variant })
 
   const attributeVariants = cva(
     'prose prose-lg dark:prose-invert prose-headings:font-title font-default focus:outline-none max-w-full h-fit prose-ol:m-0 prose-ul:m-0 prose-headings:m-0 prose-code:m-0',
@@ -75,15 +77,6 @@ const NoteEditor = ({
       }
     }
   )
-
-  const hasH1TitleEnabled = variant === 'note'
-
-  const filteredSuggestionItems = useMemo(() => {
-    if (hasH1TitleEnabled) {
-      return suggestionItems
-    }
-    return suggestionItems.filter(item => item.title !== 'Heading 1')
-  }, [hasH1TitleEnabled])
 
   return (
     <div
@@ -113,11 +106,12 @@ const NoteEditor = ({
             handleDOMEvents: {
               keydown: (_view, event) => handleCommandNavigation(event)
             },
-            handlePaste: (view, event) =>
-              handleImagePaste(view, event, uploadFn),
-            handleDrop: (view, event, _slice, moved) =>
-              handleImageDrop(view, event, moved, uploadFn),
-
+            handlePaste: (view, event) => {
+              handleImagePaste(view, event, uploadFn)
+            },
+            handleDrop: (view, event, _slice, moved) => {
+              handleImageDrop(view, event, moved, uploadFn)
+            },
             attributes: {
               class: attributeVariants({ variant })
             }
@@ -131,7 +125,7 @@ const NoteEditor = ({
               }
             }
           }}
-          editable={editable}
+          editable={editable && !isSubmitting}
           slotAfter={editable ?? <ImageResizer />}
           onCreate={({ editor }: { editor: Editor }) => {
             setEditor(editor)
