+++ src/components/editor/use-editor-state.ts
@@ -1,0 +1,44 @@
+import { type Editor } from '@tiptap/core'
+import { useEffect, useMemo, useState } from 'react'
+import { suggestionItems } from './slash-command'
+
+type UseEditorProps = {
+  isSubmitting?: boolean
+  variant?: 'note' | 'readonly' | 'videoCommentInput' | 'postInput'
+}
+const useEditorState = ({ isSubmitting, variant }: UseEditorProps) => {
+  const [openNode, setOpenNode] = useState(false)
+  const [openColor, setOpenColor] = useState(false)
+  const [openLink, setOpenLink] = useState(false)
+  const [editor, setEditor] = useState<Editor | null>(null)
+
+  useEffect(() => {
+    if (isSubmitting) {
+      editor?.commands.setContent(null)
+    }
+  }, [editor?.commands, isSubmitting])
+
+  const hasH1TitleEnabled = variant === 'note'
+
+  const filteredSuggestionItems = useMemo(() => {
+    if (hasH1TitleEnabled) {
+      return suggestionItems
+    }
+    return suggestionItems.filter(item => item.title !== 'Heading 1')
+  }, [hasH1TitleEnabled])
+
+  return {
+    openNode,
+    setOpenNode,
+    openColor,
+    setOpenColor,
+    openLink,
+    setOpenLink,
+    editor,
+    setEditor,
+    filteredSuggestionItems,
+    hasH1TitleEnabled
+  }
+}
+
+export { useEditorState }
